+++ src/components/Container.tsx
@@ -2,55 +2,23 @@
 import { Graph } from './Graph';
 import { Selector } from './Selector';
 import styles from '@/styles/components/container.module.scss';
-import { useCallback, useState } from 'react';
-import type { PopulationGraphData } from '@/interfaces/population';
-import type { MouseEventHandler } from 'react';
-import type {
-  DisplayConditions,
-  GetPrefecturesData,
-  PrefecturesList,
-} from '@/interfaces/prefectures';
+import { useDataSelector } from '@/hooks/useDataSelector';
+import type { GetPrefecturesData } from '@/interfaces/prefectures';
 
 interface Props {
   prefecturesData: GetPrefecturesData;
 }
 
 export const Contaier = ({ prefecturesData }: Props) => {
-  const [currentPrefectures, setCurrentPrefectures] =
-    useState<PrefecturesList>();
-  const [checkedPrefectures, setCheckedPrefectures] = useState<string[]>([]);
-  const [displayCondition, setDisplayCondition] =
-    useState<DisplayConditions>('総人口');
-  const [graphData, setGraphData] = useState<PopulationGraphData[]>();
-
-  const changePrefectures: MouseEventHandler<HTMLInputElement> = useCallback(
-    (e) => {
-      const currentValue = e.currentTarget.value;
-      if (!e.currentTarget.checked) {
-        setCheckedPrefectures((prev) =>
-          prev.filter((item) => !item.includes(currentValue)),
-        );
-        setGraphData(
-          (prev) => prev?.filter((item) => item.prefName !== currentValue),
-        );
-        return;
-      }
-      setCheckedPrefectures((prev) => {
-        const currentArray = [...prev, currentValue];
-        return [...new Set(currentArray)];
-      });
-      setCurrentPrefectures({
-        prefCode: Number(e.currentTarget.id),
-        prefName: currentValue,
-      });
-    },
-    [],
-  );
-
-  const changeDisplayCondition: MouseEventHandler<HTMLInputElement> =
-    useCallback((e) => {
-      setDisplayCondition(e.currentTarget.value as DisplayConditions);
-    }, []);
+  const {
+    currentPrefectures,
+    checkedPrefectures,
+    displayCondition,
+    graphData,
+    changePrefectures,
+    changeDisplayCondition,
+    setGraphData,
+  } = useDataSelector();
 
   return (
     <section className={styles.container}>
